Rename the feature one last time + improve description
